Urlencode the title to use in mailto: link
The problem was reported earlier today by Anton Yakutovich and
manifested in the crippled mailto: links - since my post's titles got
Cyrillic in them, something was going wrong on the browser side and half
the URL displayed as if it was in the wrong encoding. Anyway,
urlencoding the whole $title$ fixes the problem.