Pre-populate vim buffer with post metadata
Also start insert mode at the position where title should go. Tags field
is pre-populated by the tags that are used in existing posts. Language
field is pre-populated with space-separated list of languages I use - it
is sufficient to do dw or wdw to select one of those.

The change is sponsored by the fact that I didn't include tags in my
last three posts, so they didn't appear in runix news feed :(