Absolutize URLs in feeds
Due to the fact that all the URLs in feeds were relative to the root of
the site, FeedBurner didnt' display images - clients thought that files
are at the FeedBurner's servers!

This commit fixes this using a custom function based on the code of
`relativizeUrls`.

There is a side effect: Unicode chars and stuff like angle brackets used
to be represented like &#240; and &lt; are now used as-is. They're
inside a <![CDATA[ though, so I don't think it will cause any trouble.