Fix prev/next links for same-day posts
We maintain a chronologically ordered list of identifiers which we use
to figure out the previous and next links. The algorithm that traverses
this list made a wrong assumption that the list is also sorted
lexicographically, which broke links for posts made on the same day as
the algorithm terminated prematurely.