Fix posts page
We used to import Data.Map from containers package because that's what
Hakyll used as its Metadata type. We used the same Map from that import
to store a mapping from year to a list of posts. We were silently taking
advantage of the fact that Data.Map.keys returns keys in ordered
fashion.

When we upgraded to LTS-6, we pulled in a new version of Hakyll that
uses Aeson as its Metadata type. That, in turn, forced us to *switch*
import from Data.Map to Data.HashMap (from unordered-containers
package), as that's what Aeson uses. The Map we used to store year to
posts mapping changed accordingly, and since keys in HashMap are
unordered, the code that relied on this property of `keys` method broke.

This commit adds an explicit `sort` to fix that.