Re-design the template
I'm squashing the whole process into one commit because it was an
overhaul, not an evolution: I started pretty much from scratch and
created new template. Then I almost tweaked it to death. (Look at the
date on this commit. That's when I started working on this. Today's
November 11th, 2016, and it's only now that I'm rolling this out.)

Actually I planned to use webfonts, too, but decided against at because:

1. I have misgivings about the font choices I made a few months ago;
2. I want this work to be released at last, because it's already better
   than what I have at the moment.