Use three dashes to delineate front matter
I don't know why I started using four, but it caused problems for
Markdown syntax highlighting in Vim. I ignored the problem for years
(three and a half!), and just discovered that using three dashes fixes
it.

Previously Vim would think that the front matter is a paragraph with
a title at the end.