Deprecate "tips'n'tricks" tag
This tag was with me from the very beginning of the blog. I spent quite
a lot of time pondering what language to use for the tag: my blog was
entirely in Russian at the time, but the rest of the tags were all
English words like Linux, Debian and so on. Somehow, English won.

Now I'm deprecating it because a lot of my posts these days are exactly
tips and tricks, and this tag just doesn't make it any easier to
navigate the blog. Instead, I'm opting for more topical tags like
"programming" and "Debian".

Let's just hope no-one deep-linked to that tag's page. I'm too lazy
writing a meaningful redirect for it.